+++ drift/lib/drift.dart
@@ -2,18 +2,7 @@ export 'src/dsl/columns.dart';
 export 'src/dsl/database.dart';
 export 'src/dsl/table.dart';
 
-export 'src/query_builder/dialect.dart';
-
-export 'src/query_builder/schema/column.dart';
-export 'src/query_builder/schema/entities.dart';
-export 'src/query_builder/schema/result_set.dart';
-export 'src/query_builder/schema/table.dart';
-export 'src/query_builder/schema/view.dart';
-
-export 'src/query_builder/statements/select.dart';
-
-export 'src/query_builder/results.dart';
-export 'src/query_builder/types.dart';
+export 'src/query_builder.dart';
 
 export 'src/runtime/streams/update_rules.dart'
     hide AnyUpdateQuery, MultipleUpdateQuery, SpecificUpdateQuery;