+++ drift/lib/src/query_builder/expressions/text.dart
@@ -1,4 +1,7 @@
-part of '../query_builder.dart';
+import 'functions.dart';
+import 'expression.dart';
+import 'operators.dart';
+import 'variables.dart';
 
 /// Defines methods that operate on a column storing [String] values.
 extension StringExpressionOperators on Expression<String> {
@@ -6,63 +9,14 @@ extension StringExpressionOperators on Expression<String> {
   /// are valid and how they are interpreted, check out
   /// [this tutorial](http://www.sqlitetutorial.net/sqlite-like/).
   Expression<bool> like(String regex) {
-    return _LikeOperator(this, Variable.withString(regex));
+    return likeExp(Variable.withString(regex));
   }
 
   /// Whether this column matches the given expression. For details on what patters
   /// are valid and how they are interpreted, check out
   /// [this tutorial](http://www.sqlitetutorial.net/sqlite-like/).
   Expression<bool> likeExp(Expression<String> regex) {
-    return _LikeOperator(this, regex);
-  }
-
-  /// Matches this string against the regular expression in [regex].
-  ///
-  /// The [multiLine], [caseSensitive], [unicode] and [dotAll] parameters
-  /// correspond to the parameters on [RegExp].
-  ///
-  /// Note that this function is only available when using a `NativeDatabase`.
-  /// If you need to support the web or `moor_flutter`, consider using [like]
-  /// instead.
-  Expression<bool> regexp(
-    String regex, {
-    bool multiLine = false,
-    bool caseSensitive = true,
-    bool unicode = false,
-    bool dotAll = false,
-  }) {
-    // We have a special regexp sql function that takes a third parameter
-    // to encode flags. If the least significant bit is set, multiLine is
-    // enabled. The next three bits enable case INSENSITIVITY (it's sensitive
-    // by default), unicode and dotAll.
-    var flags = 0;
-
-    if (multiLine) {
-      flags |= 1;
-    }
-    if (!caseSensitive) {
-      flags |= 2;
-    }
-    if (unicode) {
-      flags |= 4;
-    }
-    if (dotAll) {
-      flags |= 8;
-    }
-
-    if (flags != 0) {
-      return FunctionCallExpression<bool>(
-        'regexp_moor_ffi',
-        [
-          Variable.withString(regex),
-          this,
-          Variable.withInt(flags),
-        ],
-      );
-    }
-
-    // No special flags enabled, use the regular REGEXP operator
-    return _LikeOperator(this, Variable.withString(regex), operator: 'REGEXP');
+    return BinaryExpression(this, BinaryOperator.like, regex);
   }
 
   /// Whether this expression contains [substring].
@@ -77,13 +31,13 @@ extension StringExpressionOperators on Expression<String> {
   /// Uses the given [collate] sequence when comparing this column to other
   /// values.
   Expression<String> collate(Collate collate) {
-    return _CollateOperator(this, collate);
+    return BinaryExpression(
+        this, BinaryOperator.collate, Variable.withString(collate.name));
   }
 
   /// Performs a string concatenation in sql by appending [other] to `this`.
   Expression<String> operator +(Expression<String> other) {
-    return BaseInfixOperator(this, '||', other,
-        precedence: Precedence.stringConcatenation);
+    return BinaryExpression(this, BinaryOperator.stringConcatenation, other);
   }
 
   /// Calls the sqlite function `UPPER` on `this` string. Please note that, in
@@ -137,8 +91,7 @@ extension StringExpressionOperators on Expression<String> {
   /// and [length] can be negative to return a section of the string before
   /// [start].
   Expression<String> substr(int start, [int? length]) {
-    return substrExpr(
-        Constant(start), length != null ? Constant(length) : null);
+    return substrExpr(Literal(start), length != null ? Literal(length) : null);
   }
 
   /// Calls the [`substr`](https://sqlite.org/lang_corefunc.html#substr)
@@ -155,7 +108,7 @@ extension StringExpressionOperators on Expression<String> {
   /// ```
   ///
   /// When both [start] and [length] are Dart values (e.g. [Variable]s or
-  /// [Constant]s), consider using [substr] instead.
+  /// [Literal]s), consider using [substr] instead.
   Expression<String> substrExpr(Expression<int> start,
       [Expression<int>? length]) {
     return FunctionCallExpression('SUBSTR', [
@@ -166,51 +119,11 @@ extension StringExpressionOperators on Expression<String> {
   }
 }
 
-/// A `text LIKE pattern` expression that will be true if the first expression
-/// matches the pattern given by the second expression.
-class _LikeOperator extends Expression<bool> {
-  /// The target expression that will be tested
-  final Expression<String> target;
-
-  /// The regex-like expression to test the [target] against.
-  final Expression<String> regex;
-
-  /// The operator to use when matching. Defaults to `LIKE`.
-  final String operator;
-
-  @override
-  final Precedence precedence = Precedence.comparisonEq;
-
-  /// Perform a like operator with the target and the regex.
-  _LikeOperator(this.target, this.regex, {this.operator = 'LIKE'});
-
-  @override
-  void writeInto(GenerationContext context) {
-    writeInner(context, target);
-    context.writeWhitespace();
-    context.buffer.write(operator);
-    context.writeWhitespace();
-    writeInner(context, regex);
-  }
-
-  @override
-  int get hashCode => Object.hash(target, regex, operator);
-
-  @override
-  bool operator ==(Object other) {
-    return other is _LikeOperator &&
-        other.target == target &&
-        other.regex == regex &&
-        other.operator == operator;
-  }
-}
-
 /// Collating functions used to compare texts in SQL.
 ///
 /// See also:
 /// - https://www.sqlite.org/datatype3.html#collation
-@sealed
-class Collate {
+final class Collate {
   /// The name of this collation in SQL.
   final String name;
 
@@ -231,38 +144,4 @@ class Collate {
   /// The same as [Collate.binary], except that trailing space characters are
   /// ignored.
   static const rTrim = Collate('RTRIM');
-}
-
-/// A `text COLLATE collate` expression in sqlite.
-class _CollateOperator extends Expression<String> {
-  /// The expression on which the collate function will be run
-  final Expression inner;
-
-  /// The [Collate] to use.
-  final Collate collate;
-
-  @override
-  final Precedence precedence = Precedence.postfix;
-
-  /// Constructs a collate expression on the [inner] expression and the
-  /// [Collate].
-  _CollateOperator(this.inner, this.collate);
-
-  @override
-  void writeInto(GenerationContext context) {
-    writeInner(context, inner);
-    context.buffer
-      ..write(' COLLATE ')
-      ..write(collate.name);
-  }
-
-  @override
-  int get hashCode => Object.hash(inner, collate.name);
-
-  @override
-  bool operator ==(Object other) {
-    return other is _CollateOperator &&
-        other.inner == inner &&
-        other.collate.name == collate.name;
-  }
 }